Factor out input spline support
