+++ crest/Assets/Crest/Crest/Scripts/LodData/RegisterAnimWavesInput.cs
@@ -13,13 +13,11 @@ namespace Crest
     /// Registers a custom input to the wave shape. Attach this GameObjects that you want to render into the displacmeent textures to affect ocean shape.
     /// </summary>
     [ExecuteAlways]
-    public class RegisterAnimWavesInput : RegisterLodDataInput<LodDataMgrAnimWaves>
-#if UNITY_EDITOR
-        , IReceiveSplinePointOnDrawGizmosSelectedMessages
-#endif
+    public class RegisterAnimWavesInput : RegisterLodDataInputWithSplineSupport<LodDataMgrAnimWaves>
     {
         public override bool Enabled => true;
 
+        [Header("Anim Waves Input Settings")]
         [SerializeField, Tooltip("Which octave to render into, for example set this to 2 to use render into the 2m-4m octave. These refer to the same octaves as the wave spectrum editor. Set this value to 0 to render into all LODs.")]
         float _octaveWavelength = 0f;
         public override float Wavelength
@@ -35,7 +33,9 @@ namespace Crest
 
         protected override string ShaderPrefix => "Crest/Inputs/Animated Waves";
 
-        [SerializeField, Tooltip("Whether this input data should displace horizontally with waves. If false, data will not move from side to side with the waves. Adds a small performance overhead when disabled.")]
+        protected override string SplineShaderName => "Crest/Inputs/Animated Waves/Add Water Height From Geometry";
+
+        [SerializeField, Tooltip(k_displacementCorrectionTooltip)]
         bool _followHorizontalMotion = true;
         protected override bool FollowHorizontalMotion => _followHorizontalMotion;
 
@@ -47,57 +47,9 @@ namespace Crest
         [SerializeField, Tooltip("Use the bounding box of an attached renderer component to determine the max vertical displacement.")]
         bool _reportRendererBoundsToOceanSystem = false;
 
-        [Header("Spline settings")]
-        [SerializeField]
-        float _radius = 20f;
-        [SerializeField]
-        int _subdivisions = 1;
-        [SerializeField]
-        int _smoothingIterations = 0;
-
-        Material _splineMaterial;
-        Spline.Spline _spline;
-        Mesh _splineMesh;
-
-        void Awake()
-        {
-            if (TryGetComponent<Spline.Spline>(out _spline))
-            {
-                ShapeGerstnerSplineHandling.GenerateMeshFromSpline(_spline, transform, _subdivisions, _radius, _smoothingIterations, ref _splineMesh);
-
-                if (_splineMaterial == null)
-                {
-                    _splineMaterial = new Material(Shader.Find("Crest/Inputs/Animated Waves/Add Water Height From Geometry"));
-                }
-            }
-        }
-
         protected override void Update()
         {
             base.Update();
-
-            // Check for spline and rebuild spline mesh each frame in edit mode
-            if (!EditorApplication.isPlaying)
-            {
-                if (_spline == null)
-                {
-                    TryGetComponent<Spline.Spline>(out _spline);
-                }
-
-                if (_spline != null)
-                {
-                    ShapeGerstnerSplineHandling.GenerateMeshFromSpline(_spline, transform, _subdivisions, _radius, _smoothingIterations, ref _splineMesh);
-
-                    if (_splineMaterial == null)
-                    {
-                        _splineMaterial = new Material(Shader.Find("Hidden/Crest/Inputs/Flow/Spline Geometry"));
-                    }
-                }
-                else
-                {
-                    _splineMesh = null;
-                }
-            }
 
             if (OceanRenderer.Instance == null)
             {
@@ -123,39 +75,10 @@ namespace Crest
             }
         }
 
-        public override void Draw(CommandBuffer buf, float weight, int isTransition, int lodIdx)
-        {
-            if (weight <= 0f) return;
-
-            if (_splineMesh != null && _splineMaterial != null)
-            {
-                buf.SetGlobalFloat(sp_Weight, weight);
-                buf.SetGlobalFloat(LodDataMgr.sp_LD_SliceIndex, lodIdx);
-                buf.SetGlobalVector(sp_DisplacementAtInputPosition, Vector3.zero);
-                buf.DrawMesh(_splineMesh, transform.localToWorldMatrix, _splineMaterial);
-            }
-            else
-            {
-                base.Draw(buf, weight, isTransition, lodIdx);
-            }
-        }
-
 #if UNITY_EDITOR
         // Animated waves are always enabled
-        protected override bool RendererRequired => _spline == null;
         protected override bool FeatureEnabled(OceanRenderer ocean) => true;
         protected override void FixOceanFeatureDisabled(SerializedObject oceanComponent) { }
-
-        protected new void OnDrawGizmosSelected()
-        {
-            Gizmos.color = GizmoColor;
-            Gizmos.DrawWireMesh(_splineMesh, transform.position, transform.rotation, transform.lossyScale);
-        }
-
-        public void OnSplinePointDrawGizmosSelected(SplinePoint point)
-        {
-            OnDrawGizmosSelected();
-        }
 #endif // UNITY_EDITOR
     }
 }
