+++ crest/Assets/Crest/Crest/Scripts/LodData/RegisterFlowInput.cs
@@ -13,11 +13,8 @@ namespace Crest
     /// Registers a custom input to the flow data. Attach this GameObjects that you want to influence the horizontal flow of the water volume.
     /// </summary>
     [ExecuteAlways]
-    public class RegisterFlowInput : RegisterLodDataInputDisplacementCorrection<LodDataMgrFlow>
+    public class RegisterFlowInput : RegisterLodDataInputWithSplineSupport<LodDataMgrFlow>
         , ISplinePointCustomDataSetup
-#if UNITY_EDITOR
-        , IReceiveSplinePointOnDrawGizmosSelectedMessages
-#endif
     {
         public override bool Enabled => true;
 
@@ -27,63 +24,28 @@ namespace Crest
 
         protected override string ShaderPrefix => "Crest/Inputs/Flow";
 
-        [Header("Spline settings")]
-        [SerializeField]
-        float _radius = 20f;
-        [SerializeField]
-        int _subdivisions = 1;
-        [SerializeField]
-        int _smoothingIterations = 0;
+        protected override bool FollowHorizontalMotion => _followHorizontalMotion;
+
+        protected override string SplineShaderName => "Hidden/Crest/Inputs/Flow/Spline Geometry";
+
+        [Header("Flow Input Settings")]
         [SerializeField]
         float _speed = 5f;
 
-        Material _splineMaterial;
-        Spline.Spline _spline;
-        Mesh _splineMesh;
+        [SerializeField, Tooltip(k_displacementCorrectionTooltip)]
+        bool _followHorizontalMotion = false;
 
-        void Awake()
-        {
-            if (TryGetComponent<Spline.Spline>(out _spline))
-            {
-                ShapeGerstnerSplineHandling.GenerateMeshFromSpline(_spline, transform, _subdivisions, _radius, _smoothingIterations, ref _splineMesh);
-
-                if (_splineMaterial == null)
-                {
-                    _splineMaterial = new Material(Shader.Find("Hidden/Crest/Inputs/Flow/Spline Geometry"));
-                }
-            }
-        }
-
-#if UNITY_EDITOR
         protected override void Update()
         {
             base.Update();
 
-            // Check for spline and rebuild spline mesh each frame in edit mode
-            if (!EditorApplication.isPlaying)
+            if (_splineMaterial != null)
             {
-                if (_spline == null)
-                {
-                    TryGetComponent<Spline.Spline>(out _spline);
-                }
-
-                if (_spline != null)
-                {
-                    ShapeGerstnerSplineHandling.GenerateMeshFromSpline(_spline, transform, _subdivisions, _radius, _smoothingIterations, ref _splineMesh);
-
-                    if (_splineMaterial == null)
-                    {
-                        _splineMaterial = new Material(Shader.Find("Hidden/Crest/Inputs/Flow/Spline Geometry"));
-                    }
-                }
-                else
-                {
-                    _splineMesh = null;
-                }
+                _splineMaterial.SetFloat("_Speed", _speed);
             }
         }
 
-        protected override bool RendererRequired => _spline == null;
+#if UNITY_EDITOR
         protected override bool FeatureEnabled(OceanRenderer ocean) => ocean.CreateFlowSim;
         protected override string FeatureDisabledErrorMessage => "<i>Create Flow Sim</i> must be enabled on the OceanRenderer component to enable flow on the water surface.";
         protected override void FixOceanFeatureDisabled(SerializedObject oceanComponent)
@@ -93,38 +55,7 @@ namespace Crest
 
         protected override string RequiredShaderKeyword => LodDataMgrFlow.MATERIAL_KEYWORD;
         protected override string KeywordMissingErrorMessage => LodDataMgrFlow.ERROR_MATERIAL_KEYWORD_MISSING;
-
-
-        protected new void OnDrawGizmosSelected()
-        {
-            Gizmos.color = GizmoColor;
-            Gizmos.DrawWireMesh(_splineMesh, transform.position, transform.rotation, transform.lossyScale);
-        }
-
-        public void OnSplinePointDrawGizmosSelected(SplinePoint point)
-        {
-            OnDrawGizmosSelected();
-        }
 #endif // UNITY_EDITOR
-
-        public override void Draw(CommandBuffer buf, float weight, int isTransition, int lodIdx)
-        {
-            if (weight <= 0f) return;
-
-            if (_splineMesh != null && _splineMaterial != null)
-            {
-                _splineMaterial.SetFloat("_Speed", _speed);
-
-                buf.SetGlobalFloat(sp_Weight, weight);
-                buf.SetGlobalFloat(LodDataMgr.sp_LD_SliceIndex, lodIdx);
-                buf.SetGlobalVector(sp_DisplacementAtInputPosition, Vector3.zero);
-                buf.DrawMesh(_splineMesh, transform.localToWorldMatrix, _splineMaterial);
-            }
-            else
-            {
-                base.Draw(buf, weight, isTransition, lodIdx);
-            }
-        }
 
         public bool AttachDataToSplinePoint(GameObject splinePoint)
         {
