+++ crest/Assets/Crest/Crest/Scripts/LodData/RegisterFoamInput.cs
@@ -11,7 +11,7 @@ namespace Crest
     /// Registers a custom input to the foam simulation. Attach this GameObjects that you want to influence the foam simulation, such as depositing foam on the surface.
     /// </summary>
     [ExecuteAlways]
-    public class RegisterFoamInput : RegisterLodDataInputDisplacementCorrection<LodDataMgrFoam>
+    public class RegisterFoamInput : RegisterLodDataInput<LodDataMgrFoam>
     {
         public override bool Enabled => true;
 
@@ -20,6 +20,11 @@ namespace Crest
         protected override Color GizmoColor => new Color(1f, 1f, 1f, 0.5f);
 
         protected override string ShaderPrefix => "Crest/Inputs/Foam";
+
+        protected override bool FollowHorizontalMotion => _followHorizontalMotion;
+
+        [SerializeField, Tooltip(k_displacementCorrectionTooltip)]
+        bool _followHorizontalMotion = false;
 
 #if UNITY_EDITOR
         protected override bool FeatureEnabled(OceanRenderer ocean) => ocean.CreateFoamSim;
